pyproject.toml
[project]
name = "fastapi-async-auth-kit"
version = "0.3.0"
description = "Production-ready FastAPI auth system with JWT, refresh tokens, and async DB support"
readme = "README.md"
authors = [{ name="Keyurkumar", email="kmistry1110@gmail.com" }]
license = "MIT"
requires-python = ">=3.8"

dependencies = [
    "fastapi",
    "pydantic",
    "python-jose",
    "argon2-cffi",
    "sqlalchemy>=2.0"
]

[project.optional-dependencies]
postgres = ["asyncpg"]
mysql = ["aiomysql"]
sqlite = ["aiosqlite"]
mongodb = ["motor"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["fastapi_async_auth_kit*"]